pyproject.toml
[tool.poetry]
name = "split-cost"
version = "0.1.0"
description = ""
authors = ["cgtyunerr <cgtyuner06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
uvicorn = "^0.30.5"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
orjson = "^3.10.6"
bcrypt = "4.0.1"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
